Fix unmanaged in the efinoruntime project
Roslyn is very particular about the shape of the `RuntimeFeature` class: it must be both in the assembly that defines System.Object _and_ be a `static` class. Just a class doesn't cut it. I also added LangVersion 9 to the project file to get rid of the langversion error when opening the csproj in VS or VSCode.